refactor: extract out types
